DataFlash: Change ::bufferspace_available return type...
from 'uint16_t' to 'uint32_t' to avoid overflow (possible in at least
two implementations: Block and File).

+++ libraries/DataFlash/DataFlash_Backend.h
@@ -60,7 +60,7 @@ public:
 
     void set_mission(const AP_Mission *mission);
 
-    virtual uint16_t bufferspace_available() = 0;
+    virtual uint32_t bufferspace_available() = 0;
 
     virtual uint16_t start_new_log(void) = 0;
     bool log_write_started;
